+++ CHANGELOG.md
@@ -21,6 +21,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   finance team
 - Added a field to enter the date a grant payment certificate was received to
   the Conversions Receive grant payment certificate task.
+- Script to backfill dates in the Conversions Receive grant payment certificate
+  task
 
 ### Changed